Fixed compilation for windows 7 in 32 bits.
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@738 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/Language/VarContexts.hpp
@@ -94,7 +94,7 @@ public:
    * @param  
    */
   void SetCurrentContextNumber( int context) {
-    if ((context>=0)&&(context<_context.size()))
+    if ((context>=0)&&(context<(int)_context.size()))
       _current_context = context;
   }
